Converted react-scripts to vite
In DEV mode, you just do:
go run .
(cd ui && yarn start)
In dev mode there is a proxy in vite used to not have CORS problem.
You can access the UI on:
http://localhost:5173
The UI prefix every request to the backup with `/api` so the proxy can correctly handle it, the proxy strips the `/api` so the backend is equal to production mode.